+++ HISTORY.md
@@ -3,7 +3,14 @@
 
 # Release history
 
-v1.30.0
+v1.31.0 ()
+------------------
+
+* Support online check ip which used by wwanleds/cloud_watch service
+* Support data cache enable command
+
+
+v1.30.0 (2021-04-22)
 ------------------
 
 * Improved lockable_queue